fix(core): prevent better-sqlite3 import leaking into Postgres server bundle
